This commit is much bigger than I typically like.  Way too many changes.  I'll try to comment on as many changes as possible.
Added a method to handle the adjustment of the webview.

Added a method to handle setting and adjusting the scrollview content size.

Rather than defining a fixed value, using the view bounds for things like size height or size width is used to help make things more dynamic.

Added flag property to indicate when content is loaded in the webview.  Make use of it in the add/remove content methods.

Changed the order in which removeContent adds empty content and sets the height for the webview.

Refactored adjustHeight:forView: a bit to handle orientation changes better. Also set the heightLabel here now, as this is where it's actually determined.

Moved the javascript code and call on the webview for determining height on the webview to it's own method.  It's used in a couple of places now.

Commented out code inside of willRotateToInterface of another attempt at trying to solve the viewport problem.  Still think it's a viable approach.  Damn javascript does not execute correctly for some reason.

didRotateFromInterface method implemented to make sure we adjust the webview when we need to.

Commented out method for viewDidLayoutSubviews included.  It's a reminder of a failed path to solving this problem.

Removed some code that was not being used that was copied in from an example.

Added some comments.